Split auth'ed routes, move Sidebar to layout
