+++ apps/app/app/(authenticated)/page.tsx
@@ -1,0 +1,23 @@
+import type { Metadata } from 'next';
+import type { FC } from 'react';
+
+const title = 'Acme Inc';
+const description = 'My application.';
+
+export const metadata: Metadata = {
+  title,
+  description,
+};
+
+const App: FC = () => (
+  <div className="flex flex-1 flex-col gap-4 p-4 pt-0">
+    <div className="grid auto-rows-min gap-4 md:grid-cols-3">
+      <div className="aspect-video rounded-xl bg-muted/50" />
+      <div className="aspect-video rounded-xl bg-muted/50" />
+      <div className="aspect-video rounded-xl bg-muted/50" />
+    </div>
+    <div className="min-h-[100vh] flex-1 rounded-xl bg-muted/50 md:min-h-min" />
+  </div>
+);
+
+export default App;
